HUD refactoring (part 2)

+++ implementation/src/main/java/br/com/guigasgame/team/HeroTeam.java
@@ -112,7 +112,7 @@ public class HeroTeam
 		return enemiesMask;
 	}
 
-	public float getTeamId()
+	public int getTeamId()
 	{
 		return teamId;
 	}
